+++ pytorch_forecasting/data/timeseries/__init__.py
@@ -1,0 +1,9 @@
+"""Data loaders for time series data."""
+
+from pytorch_forecasting.data.timeseries._timeseries_v2 import TimeSeries
+from pytorch_forecasting.data.timeseries._timeseries import TimeSeriesDataSet
+
+__all__ = [
+    "TimeSeriesDataSet",
+    "TimeSeries",
+]
